test: 구입 금액 검증 로직 테스트 코드 추가
